revert 0 length throw, and update sequence fuzz
