+++ packages/dds/merge-tree/src/test/beastTest.spec.ts
@@ -741,11 +741,13 @@ export function TestPack(verbose = true) {
 	function clientServer(startFile?: string, initRounds = 1000) {
 		const clientCount = 5;
 		const fileSegCount = 0;
-		const server = new TestServer();
-		const initString = !startFile ? "don't ask for whom the bell tolls; it tolls for thee" : "";
-		if (initString.length > 0) {
-			server.insertTextLocal(0, initString);
+		let initString = "";
+
+		if (!startFile) {
+			initString = "don't ask for whom the bell tolls; it tolls for thee";
 		}
+		const server = new TestServer();
+		server.insertTextLocal(0, initString);
 		server.measureOps = true;
 		if (startFile) {
 			loadTextFromFile(startFile, server.mergeTree, fileSegCount);
@@ -754,9 +756,7 @@ export function TestPack(verbose = true) {
 		const clients = new Array<TestClient>(clientCount);
 		for (let i = 0; i < clientCount; i++) {
 			clients[i] = new TestClient();
-			if (initString.length > 0) {
-				clients[i].insertTextLocal(0, initString);
-			}
+			clients[i].insertTextLocal(0, initString);
 			clients[i].measureOps = true;
 			if (startFile) {
 				loadTextFromFile(startFile, clients[i].mergeTree, fileSegCount);
