+++ packages/dds/sequence/src/test/fuzz/fuzzUtils.ts
@@ -6,8 +6,10 @@ import * as path from "path";
 import { strict as assert } from "assert";
 import {
 	AcceptanceCondition,
-	combineReducersAsync as combineReducers,
-	AsyncReducer as Reducer,
+	combineReducersAsync,
+	AsyncReducer,
+	AsyncGenerator,
+	createWeightedAsyncGenerator,
 } from "@fluid-private/stochastic-test-utils";
 import { DDSFuzzModel, DDSFuzzSuiteOptions, DDSFuzzTestState } from "@fluid-private/test-dds-utils";
 import {
@@ -19,8 +21,10 @@ import { PropertySet } from "@fluidframework/merge-tree";
 import { revertSharedStringRevertibles, SharedStringRevertible } from "../../revertibles.js";
 import { SharedStringFactory } from "../../sequenceFactory.js";
 import { SharedString } from "../../sharedString.js";
-import { Side } from "../../intervalCollection.js";
+import { Side, type IIntervalCollection } from "../../intervalCollection.js";
 import { assertEquivalentSharedStrings } from "../intervalTestUtils.js";
+import { _dirname } from "../dirname.cjs";
+import type { SequenceInterval } from "../../index.js";
 
 export type RevertibleSharedString = SharedString & {
 	revertibles: SharedStringRevertible[];
@@ -197,9 +201,9 @@ type ClientOpState = FuzzTestState;
 
 export function makeReducer(
 	loggingInfo?: LoggingInfo,
-): Reducer<Operation | RevertOperation, ClientOpState> {
+): AsyncReducer<Operation | RevertOperation, ClientOpState> {
 	const withLogging =
-		<T>(baseReducer: Reducer<T, ClientOpState>): Reducer<T, ClientOpState> =>
+		<T>(baseReducer: AsyncReducer<T, ClientOpState>): AsyncReducer<T, ClientOpState> =>
 		async (state, operation) => {
 			if (loggingInfo !== undefined) {
 				logCurrentState(state, loggingInfo);
@@ -209,7 +213,7 @@ export function makeReducer(
 			await baseReducer(state, operation);
 		};
 
-	const reducer = combineReducers<Operation | RevertOperation, ClientOpState>({
+	const reducer = combineReducersAsync<Operation | RevertOperation, ClientOpState>({
 		addText: async ({ client }, { index, content }) => {
 			client.channel.insertText(index, content);
 		},
@@ -406,7 +410,7 @@ export const defaultFuzzOptions: Partial<DDSFuzzSuiteOptions> = {
 		clientAddProbability: 0.1,
 	},
 	defaultTestCount: 100,
-	saveFailures: { directory: path.join(__dirname, "../../../src/test/fuzz/results") },
+	saveFailures: { directory: path.join(_dirname, "../../src/test/fuzz/results") },
 	parseOperations: (serialized: string) => {
 		const operations: Operation[] = JSON.parse(serialized);
 		// Replace this value with some other interval ID and uncomment to filter replay of the test
@@ -420,3 +424,157 @@ export const defaultFuzzOptions: Partial<DDSFuzzSuiteOptions> = {
 		return operations;
 	},
 };
+
+export function makeIntervalOperationGenerator(
+	optionsParam?: IntervalOperationGenerationConfig,
+	alwaysLeaveChar: boolean = false,
+): AsyncGenerator<Operation, ClientOpState> {
+	const {
+		startPosition,
+		addText,
+		obliterateRange,
+		removeRange,
+		removeRangeLeaveChar,
+		lengthSatisfies,
+		hasNonzeroLength,
+		isShorterThanMaxLength,
+	} = createSharedStringGeneratorOperations(optionsParam);
+
+	const options = { ...defaultIntervalOperationGenerationConfig, ...(optionsParam ?? {}) };
+
+	function isNonEmpty(collection: IIntervalCollection<SequenceInterval>): boolean {
+		for (const _ of collection) {
+			return true;
+		}
+
+		return false;
+	}
+
+	function inclusiveRange(state: ClientOpState): RangeSpec {
+		const start = startPosition(state);
+		const end = state.random.integer(
+			start,
+			Math.max(start, state.client.channel.getLength() - 1),
+		);
+		return { start, end };
+	}
+
+	function inclusiveRangeWithUndefined(
+		state: ClientOpState,
+	): RangeSpec | { start: undefined; end: undefined } {
+		return state.random.bool() ? inclusiveRange(state) : { start: undefined, end: undefined };
+	}
+
+	function propertySet(state: ClientOpState): PropertySet {
+		const propNamesShuffled = [...options.propertyNamePool];
+		state.random.shuffle(propNamesShuffled);
+		const propsToChange = propNamesShuffled.slice(
+			0,
+			state.random.integer(1, propNamesShuffled.length),
+		);
+		const propSet: PropertySet = {};
+		for (const name of propsToChange) {
+			propSet[name] = state.random.string(5);
+		}
+		return propSet;
+	}
+
+	function propertySetWithUndefined(state: ClientOpState): PropertySet | undefined {
+		return state.random.bool() ? propertySet(state) : undefined;
+	}
+
+	function nonEmptyIntervalCollection({ client, random }: ClientOpState): string {
+		const nonEmptyLabels = Array.from(client.channel.getIntervalCollectionLabels()).filter(
+			(label) => {
+				const collection = client.channel.getIntervalCollection(label);
+				return isNonEmpty(collection);
+			},
+		);
+		return random.pick(nonEmptyLabels);
+	}
+
+	function interval(state: ClientOpState): { collectionName: string; id: string } {
+		const collectionName = nonEmptyIntervalCollection(state);
+		const intervals = Array.from(state.client.channel.getIntervalCollection(collectionName));
+		const id = state.random.pick(intervals)?.getIntervalId();
+		assert(id);
+
+		return {
+			id,
+			collectionName,
+		};
+	}
+
+	async function addInterval(state: ClientOpState): Promise<AddInterval> {
+		return {
+			type: "addInterval",
+			...inclusiveRange(state),
+			collectionName: state.random.pick(options.intervalCollectionNamePool),
+			id: state.random.uuid4(),
+			startSide: state.random.pick([Side.Before, Side.After]),
+			endSide: state.random.pick([Side.Before, Side.After]),
+		};
+	}
+
+	async function deleteInterval(state: ClientOpState): Promise<DeleteInterval> {
+		return {
+			type: "deleteInterval",
+			...interval(state),
+		};
+	}
+
+	async function changeInterval(state: ClientOpState): Promise<ChangeInterval> {
+		const { start, end } = inclusiveRangeWithUndefined(state);
+		const properties = propertySetWithUndefined(state);
+		return {
+			type: "changeInterval",
+			start,
+			end,
+			startSide: state.random.pick([Side.Before, Side.After]),
+			endSide: state.random.pick([Side.Before, Side.After]),
+			properties,
+			...interval(state),
+		};
+	}
+
+	const hasAnInterval = ({ client }: ClientOpState): boolean =>
+		Array.from(client.channel.getIntervalCollectionLabels()).some((label) => {
+			const collection = client.channel.getIntervalCollection(label);
+			return isNonEmpty(collection);
+		});
+
+	const hasNotTooManyIntervals: AcceptanceCondition<ClientOpState> = ({ client }) => {
+		let intervalCount = 0;
+		for (const label of client.channel.getIntervalCollectionLabels()) {
+			for (const _ of client.channel.getIntervalCollection(label)) {
+				intervalCount++;
+				if (intervalCount >= options.maxIntervals) {
+					return false;
+				}
+			}
+		}
+		return true;
+	};
+
+	const all =
+		<T>(...clauses: AcceptanceCondition<T>[]): AcceptanceCondition<T> =>
+		(t: T) =>
+			clauses.reduce<boolean>((prev, cond) => prev && cond(t), true);
+	const usableWeights = optionsParam?.weights ?? defaultIntervalOperationGenerationConfig.weights;
+	return createWeightedAsyncGenerator<Operation, ClientOpState>([
+		[addText, usableWeights.addText, isShorterThanMaxLength],
+		[
+			alwaysLeaveChar ? removeRangeLeaveChar : removeRange,
+			usableWeights.removeRange,
+			alwaysLeaveChar
+				? lengthSatisfies((length) => {
+						return length > 1;
+				  })
+				: hasNonzeroLength,
+		],
+		[obliterateRange, usableWeights.obliterateRange, hasNonzeroLength],
+		[addInterval, usableWeights.addInterval, all(hasNotTooManyIntervals, hasNonzeroLength)],
+		[deleteInterval, usableWeights.deleteInterval, hasAnInterval],
+		[changeInterval, usableWeights.changeInterval, all(hasAnInterval, hasNonzeroLength)],
+	]);
+}
